corpus/pipeline/requirements.txt
anthropic>=0.45.0
openai>=1.60.0
python-dotenv>=1.0.0
tqdm>=4.66.0
